wj/requirements.txt
# Base
torch>=2.0
torchvision>=0.15
torchaudio>=2.0
numpy>=1.23
scipy>=1.10
scikit-learn>=1.3
scikit-image>=0.21
matplotlib>=3.7
seaborn>=0.12
opencv-python>=4.7
Pillow>=9.4
pandas>=1.5
pyyaml>=6.0
tqdm>=4.65
einops>=0.6
ftfy>=6.1
regex>=2023.6
timm>=0.9
huggingface-hub>=0.14
sentencepiece>=0.1.99
open_clip-torch

# Evaluation / profiling
loguru>=0.7
fvcore>=0.1.5.post20221221
ptflops>=0.7

# Visualization utilities
plotly>=5.15

# Optional: Hydra-style config (commented out)
# hydra-core>=1.3

